fixed isues according to teacher's feedback - changed one-way code to two-way (added more link to go back) - member cannot delete admin - user cannot delete themselves - refresh the search book grid after adding books to cart - ...

+++ MVC2SE1614_1/src/java/hadn/servlet/DeleteAccountServlet.java
@@ -6,6 +6,7 @@
 package hadn.servlet;
 
 import hadn.registration.RegistrationDAO;
+import hadn.registration.RegistrationDTO;
 import hadn.utils.MyApplicationConstants;
 import java.io.IOException;
 import java.io.PrintWriter;
@@ -20,6 +21,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 /**
  *
@@ -29,8 +31,6 @@ import javax.servlet.http.HttpServletResponse;
 public class DeleteAccountServlet extends HttpServlet {
 
 //    private final String ERROR_PAGE = "blinkError.html";
-    
-
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
      * methods.
@@ -43,34 +43,52 @@ public class DeleteAccountServlet extends HttpServlet {
     protected void processRequest(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
         ServletContext context = this.getServletContext();
-        Properties siteMaps = (Properties)context.getAttribute("SITEMAPS");
+        Properties siteMaps = (Properties) context.getAttribute("SITEMAPS");
         response.setContentType("text/html;charset=UTF-8");
         String username = request.getParameter("pk");
+        String role = request.getParameter("role");
+        boolean isAdmin = false;
+        if (role.equals("true")) {
+            isAdmin = true;
+        }
         String searchValue = request.getParameter("lastSearchValue");
 //        String url = ERROR_PAGE;
         String url = siteMaps.getProperty(MyApplicationConstants.DeleteAccountFeature.BLINK_ERROR_PAGE);
         try {
-            
-            //Practice
             //1. Call Dao
-            RegistrationDAO dao = new RegistrationDAO();
-            boolean result = dao.deleteAccount(username);
-            if (result) { //gọi search lại 1 lần nữa
-                url = "searchLastnameController?"
-//                        + "?btAction=Search"
-                        + "&txtSearchValue=" + searchValue;
+            HttpSession session = request.getSession(false);
+            if (session != null) {
+                RegistrationDTO user = (RegistrationDTO) session.getAttribute("USER");
+                if (!user.getUsername().equals(username)) {
+                    if (user.isRole() == false) {
+                        if (isAdmin == false) {
+                            RegistrationDAO dao = new RegistrationDAO();
+                            boolean result = dao.deleteAccount(username);
+                            if (result) { //gọi search lại 1 lần nữa
+                                url = "searchLastnameController?"
+                                        //                        + "?btAction=Search"
+                                        + "&txtSearchValue=" + searchValue;
+                            }
+                        } else {
+                            url = "searchLastnameController?"
+                                    + "&adminDeleteError=" + "Normal user cannot delete admin"
+                                    + "&txtSearchValue=" + searchValue;
+                        }
+                    } else {
+                        RegistrationDAO dao = new RegistrationDAO();
+                        boolean result = dao.deleteAccount(username);
+                        if (result) { //gọi search lại 1 lần nữa
+                            url = "searchLastnameController?"
+                                    //                        + "?btAction=Search"
+                                    + "&txtSearchValue=" + searchValue;
+                        }
+                    }
+                } else {
+                    url = "searchLastnameController?"
+                            + "&selfDeleteError=" + "User cannot delete themselves"
+                            + "&txtSearchValue=" + searchValue;
+                }
             }
-            //practice
-            //start
-//            RegistrationDAO dao = new RegistrationDAO();
-//            boolean result = dao.deleteAccount(username);
-//            if (result) {
-//                url = "searchLastnameController?" +
-//                        "txtSearchValue=" + searchValue;
-//            }
-            
-            
-            //end
         } catch (NamingException ex) {
             log("DeleteAccountServlet _ Naming" + ex.getMessage());
         } catch (SQLException ex) {
